requirements.txt
pypdf>=3.0.0
google-genai>=0.1.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
requests>=2.31.0
